+++ app/lib/ui/components/components_screen.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:ods_flutter/components/card/ods_small_card.dart';
-import 'package:ods_flutter/guidelines/spacings.dart';
 import 'package:ods_flutter_demo/ui/components/component_entities.dart';
 import 'package:ods_flutter_demo/ui/components/detail_screen.dart';
 import 'package:ods_flutter_demo/ui/utilities/display_image.dart';
@@ -18,16 +17,17 @@ class ComponentsScreen extends StatefulWidget {
 class _ComponentsScreenState extends State<ComponentsScreen> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.only(
-              left: spacingXs, right: spacingXs, top: spacingS),
-          child: GridView.count(
-            crossAxisCount: 2,
-            children: widget
-                .odsComponents // Access the list using widget.odsComponents
-                .map((component) => OdsSmallCard(
+    return SafeArea(
+      child: OrientationBuilder(
+        builder: (context, orientation) {
+          return GridView.count(
+            crossAxisCount: orientation == Orientation.portrait ? 2 : 3,
+            childAspectRatio: orientation == Orientation.portrait ? 1.0 : 1.2,
+            children: widget.odsComponents.map(
+              (component) {
+                return Column(
+                  children: [
+                    OdsSmallCard(
                       title: component.title,
                       image: displayImage(
                           component.imageResourceName, double.infinity),
@@ -38,10 +38,13 @@ class _ComponentsScreenState extends State<ComponentsScreen> {
                             ),
                             transition: Transition.rightToLeft);
                       },
-                    ))
-                .toList(),
-          ),
-        ),
+                    ),
+                  ],
+                );
+              },
+            ).toList(),
+          );
+        },
       ),
     );
   }