+++ library/lib/components/card/ods_small_card.dart
@@ -42,37 +42,42 @@ class _OdsSmallCardState extends State<OdsSmallCard> {
     return Stack(children: [
       SizedBox(
         width: double.infinity,
-        child: Card(
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(odsCardRadius),
-          ),
-          elevation: 1,
-          clipBehavior: Clip.antiAliasWithSaveLayer,
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              SizedBox(
-                width: double.infinity,
-                height: OdsSmallCard._imageHeight,
-                child: widget.image,
-              ),
-              Padding(
-                padding: const EdgeInsets.all(spacingM),
-                child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-                  Text(
-                    widget.title,
-                    style: Theme.of(context).textTheme.titleLarge,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                  if (widget.subtitle != null)
-                    Text(
-                      widget.subtitle!,
-                      style: Theme.of(context).textTheme.titleMedium,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                ]),
-              )
-            ],
+        child: Padding(
+          padding: const EdgeInsets.all(spacingXs),
+          child: Card(
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(odsCardRadius),
+            ),
+            elevation: 3,
+            clipBehavior: Clip.antiAliasWithSaveLayer,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                SizedBox(
+                  width: double.infinity,
+                  height: OdsSmallCard._imageHeight,
+                  child: widget.image,
+                ),
+                Padding(
+                  padding: const EdgeInsets.all(spacingM),
+                  child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          widget.title,
+                          style: Theme.of(context).textTheme.titleMedium,
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                        if (widget.subtitle != null)
+                          Text(
+                            widget.subtitle!,
+                            style: Theme.of(context).textTheme.bodyMedium,
+                            overflow: TextOverflow.ellipsis,
+                          ),
+                      ]),
+                )
+              ],
+            ),
           ),
         ),
       ),